Serialise runs of the tag and release workflow
Concurrent runs of this workflow can fail due to conflicting changes in tags.
The GitHub workflow concurrency option works to some degree, but only a
single job can be queued concurrently, and older queued jobs get
cancelled. This makes sense for a deployment, but not for a tagging
workflow.

This change uses the softprops/turnstyle action, which polls the GitHub
API and waits until other workflows have completed.

Consumers of the workflow now require the "actions: read" permission.